refactor: splitting app and customer select
